+++ test/aggregator_test.rb
@@ -1,9 +1,28 @@
 # frozen_string_literal: true
 
 require "test_helper"
+require "ostruct"
 
 class AggregatorTest < Minitest::Test
+  class CaptureLogger
+    attr_reader :messages
+
+    def initialize
+      @messages = []
+    end
+
+    [:debug, :info, :warn, :error, :fatal].each do |severity|
+      define_method(severity) do |message = nil, &block|
+        message = block.call if message.nil? && block
+        @messages << { severity: severity, message: message }
+      end
+    end
+  end
+
   def setup
+    @logger = CaptureLogger.new
+    @old_logger = StatsD.logger
+    StatsD.logger = @logger
     @sink = StatsD::Instrument::CaptureSink.new(parent: StatsD::Instrument::NullSink.new)
     @subject = StatsD::Instrument::Aggregator.new(
       @sink, StatsD::Instrument::DatagramBuilder, nil, [], flush_interval: 0.1
@@ -12,6 +31,7 @@ class AggregatorTest < Minitest::Test
 
   def teardown
     @sink.clear
+    StatsD.logger = @old_logger
   end
 
   def test_increment_simple
@@ -138,5 +158,90 @@ class AggregatorTest < Minitest::Test
 
     assert_equal("foo", @sink.datagrams.last.name)
     assert_equal(1, @sink.datagrams.last.value)
+  end
+
+  def test_synchronous_operation_on_thread_failure
+    # Force thread_healthcheck to return false
+    @subject.stubs(:thread_healthcheck).returns(false)
+
+    @subject.increment("foo", 1, tags: { foo: "bar" })
+    @subject.aggregate_timing("bar", 100, tags: { foo: "bar" })
+    @subject.gauge("baz", 100, tags: { foo: "bar" })
+
+    # Verify metrics were sent immediately
+    assert_equal(3, @sink.datagrams.size)
+
+    counter_datagram = @sink.datagrams.find { |d| d.name == "foo" }
+    assert_equal(1, counter_datagram.value)
+    assert_equal(["foo:bar"], counter_datagram.tags)
+
+    timing_datagram = @sink.datagrams.find { |d| d.name == "bar" }
+    assert_equal([100.0], [timing_datagram.value])
+    assert_equal(["foo:bar"], timing_datagram.tags)
+
+    gauge_datagram = @sink.datagrams.find { |d| d.name == "baz" }
+    assert_equal(100, gauge_datagram.value)
+    assert_equal(["foo:bar"], gauge_datagram.tags)
+
+    # Additional metrics should also go through synchronously
+    @subject.increment("foo", 1, tags: { foo: "bar" })
+    @subject.aggregate_timing("bar", 200, tags: { foo: "bar" })
+
+    # Verify new metrics were also sent immediately
+    assert_equal(5, @sink.datagrams.size)
+
+    counter_datagram = @sink.datagrams.select { |d| d.name == "foo" }.last
+    assert_equal(1, counter_datagram.value)
+    assert_equal(["foo:bar"], counter_datagram.tags)
+
+    timing_datagram = @sink.datagrams.select { |d| d.name == "bar" }.last
+    assert_equal([200.0], [timing_datagram.value])
+    assert_equal(["foo:bar"], timing_datagram.tags)
+
+    # undo the stubbing
+    @subject.unstub(:thread_healthcheck)
+  end
+
+  def test_recreate_thread_after_fork
+    skip("Not supported in non MRI platforms") unless RUBY_PLATFORM == "ruby"
+    # Record initial metrics
+    @subject.increment("foo", 1, tags: { foo: "bar" })
+    @subject.aggregate_timing("bar", 100, tags: { foo: "bar" })
+
+    # kill the flush thread
+    @subject.instance_variable_get(:@flush_thread).kill
+
+    # Fork the process
+    pid = Process.fork do
+      # In forked process, send more metrics
+      @subject.increment("foo", 2, tags: { foo: "bar" })
+      @subject.aggregate_timing("bar", 200, tags: { foo: "bar" })
+      @subject.flush
+      assert_equal(4, @sink.datagrams.size)
+      exit!
+    end
+
+    @logger.messages.each { |m| $stdout.puts m[:message] }
+
+    # Wait for forked process to complete
+    Process.wait(pid)
+
+    # Send metrics in parent process
+    @subject.increment("foo", 3, tags: { foo: "bar" })
+    @subject.aggregate_timing("bar", 300, tags: { foo: "bar" })
+    @subject.flush
+
+    assert_equal(2, @sink.datagrams.size)
+
+    # Verify metrics were properly aggregated in parent process
+    counter_datagrams = @sink.datagrams.select { |d| d.name == "foo" }
+    timing_datagrams = @sink.datagrams.select { |d| d.name == "bar" }
+
+    assert_equal(1, counter_datagrams.size)
+    assert_equal(1, timing_datagrams.size)
+
+    # Aggregate despite fork
+    assert_equal(4, counter_datagrams.last.value)
+    assert_equal([100.0, 300.0], timing_datagrams.last.value)
   end
 end